backend/requirements.txt
Django>=4.2
djangorestframework
dj-database-url
psycopg2-binary
python-dotenv
django-cors-headers
cloudinary
django-cloudinary-storage
drf-spectacular
resend
requests
Pillow
